requirements.txt
crewai>=0.35.0
langchain-openai>=0.1.7
python-dotenv>=1.0.0
pydantic>=2.0
